Move webhook errors back to models
