+++ models/webhook/webhook.go
@@ -21,6 +21,46 @@ import (
 	"xorm.io/builder"
 )
 
+// ErrWebhookNotExist represents a "WebhookNotExist" kind of error.
+type ErrWebhookNotExist struct {
+	ID int64
+}
+
+// IsErrWebhookNotExist checks if an error is a ErrWebhookNotExist.
+func IsErrWebhookNotExist(err error) bool {
+	_, ok := err.(ErrWebhookNotExist)
+	return ok
+}
+
+func (err ErrWebhookNotExist) Error() string {
+	return fmt.Sprintf("webhook does not exist [id: %d]", err.ID)
+}
+
+func (err ErrWebhookNotExist) Unwrap() error {
+	return util.ErrNotExist
+}
+
+// ErrHookTaskNotExist represents a "HookTaskNotExist" kind of error.
+type ErrHookTaskNotExist struct {
+	TaskID int64
+	HookID int64
+	UUID   string
+}
+
+// IsErrHookTaskNotExist checks if an error is a ErrHookTaskNotExist.
+func IsErrHookTaskNotExist(err error) bool {
+	_, ok := err.(ErrHookTaskNotExist)
+	return ok
+}
+
+func (err ErrHookTaskNotExist) Error() string {
+	return fmt.Sprintf("hook task does not exist [task: %d, hook: %d, uuid: %s]", err.TaskID, err.HookID, err.UUID)
+}
+
+func (err ErrHookTaskNotExist) Unwrap() error {
+	return util.ErrNotExist
+}
+
 // HookContentType is the content type of a web hook
 type HookContentType int
 
@@ -359,7 +399,7 @@ func getWebhook(bean *Webhook) (*Webhook, error) {
 	if err != nil {
 		return nil, err
 	} else if !has {
-		return nil, webhook_module.ErrWebhookNotExist{ID: bean.ID}
+		return nil, ErrWebhookNotExist{ID: bean.ID}
 	}
 	return bean, nil
 }
@@ -447,7 +487,7 @@ func GetSystemOrDefaultWebhook(id int64) (*Webhook, error) {
 	if err != nil {
 		return nil, err
 	} else if !has {
-		return nil, webhook_module.ErrWebhookNotExist{ID: id}
+		return nil, ErrWebhookNotExist{ID: id}
 	}
 	return webhook, nil
 }
@@ -489,7 +529,7 @@ func deleteWebhook(bean *Webhook) (err error) {
 	if count, err := db.DeleteByBean(ctx, bean); err != nil {
 		return err
 	} else if count == 0 {
-		return webhook_module.ErrWebhookNotExist{ID: bean.ID}
+		return ErrWebhookNotExist{ID: bean.ID}
 	} else if _, err = db.DeleteByBean(ctx, &HookTask{HookID: bean.ID}); err != nil {
 		return err
 	}
@@ -527,7 +567,7 @@ func DeleteDefaultSystemWebhook(id int64) error {
 	if err != nil {
 		return err
 	} else if count == 0 {
-		return webhook_module.ErrWebhookNotExist{ID: id}
+		return ErrWebhookNotExist{ID: id}
 	}
 
 	if _, err := db.DeleteByBean(ctx, &HookTask{HookID: id}); err != nil {
