+++ routers/web/repo/webhook.go
@@ -594,7 +594,7 @@ func checkWebhook(ctx *context.Context) (*orgRepoCtx, *webhook.Webhook) {
 		w, err = webhook.GetSystemOrDefaultWebhook(ctx.ParamsInt64(":id"))
 	}
 	if err != nil || w == nil {
-		if webhook_module.IsErrWebhookNotExist(err) {
+		if webhook.IsErrWebhookNotExist(err) {
 			ctx.NotFound("GetWebhookByID", nil)
 		} else {
 			ctx.ServerError("GetWebhookByID", err)
@@ -708,7 +708,7 @@ func ReplayWebhook(ctx *context.Context) {
 	}
 
 	if err := webhook_service.ReplayHookTask(ctx, w, hookTaskUUID); err != nil {
-		if webhook_module.IsErrHookTaskNotExist(err) {
+		if webhook.IsErrHookTaskNotExist(err) {
 			ctx.NotFound("ReplayHookTask", nil)
 		} else {
 			ctx.ServerError("ReplayHookTask", err)
